+++ src/languages/system.rs
@@ -1,8 +1,10 @@
+use std::sync::Arc;
 use tokio::process::Command;
 
 use crate::config;
 use crate::hook::Hook;
 use crate::languages::{LanguageImpl, DEFAULT_VERSION};
+use crate::run::run_by_batch;
 
 #[derive(Debug, Copy, Clone)]
 pub struct System;
@@ -30,13 +32,39 @@ impl LanguageImpl for System {
 
     async fn run(&self, hook: &Hook, filenames: &[&String]) -> anyhow::Result<(i32, Vec<u8>)> {
         let cmds = shlex::split(&hook.entry).ok_or(anyhow::anyhow!("Failed to parse entry"))?;
-        let output = Command::new(&cmds[0])
-            .args(&cmds[1..])
-            .args(&hook.args)
-            .args(filenames)
-            .stderr(std::process::Stdio::inherit())
-            .output()
-            .await?;
-        Ok((output.status.code().unwrap_or(1), output.stdout))
+
+        let cmds = Arc::new(cmds);
+        let hook_args = Arc::new(hook.args.clone());
+
+        let run = move |batch: Vec<String>| {
+            let cmds = cmds.clone();
+            let hook_args = hook_args.clone();
+
+            async move {
+                let output = Command::new(&cmds[0])
+                    .args(&cmds[1..])
+                    .args(hook_args.as_ref())
+                    .args(batch)
+                    .stderr(std::process::Stdio::inherit())
+                    .output()
+                    .await?;
+
+                let code = output.status.code().unwrap_or(1);
+                anyhow::Ok((code, output.stdout))
+            }
+        };
+
+        let results = run_by_batch(hook, filenames, run).await?;
+
+        // Collect results
+        let mut combined_status = 0;
+        let mut combined_output = Vec::new();
+
+        for (code, output) in results {
+            combined_status |= code;
+            combined_output.extend(output);
+        }
+
+        Ok((combined_status, combined_output))
     }
 }
